add all test

+++ services/UserService.js
@@ -6,42 +6,60 @@ const ObjectId = mongoose.Types.ObjectId
 
 var User = mongoose.model('User', UserSchema)
 
-module.exports.addOneUser = function (user, callback) {
-    const new_user = new User(user)
-    let errors = new_user.validateSync()
-    if (errors) {
-        errors = errors['errors']
-        var text = Object.keys(errors).map((e) => {
-            return errors[e]['properties']['message']
-        }).join(' ')
-        var fields = _.transform(Object.keys(errors), function (result, value) {
-            result[value] = errors[value]['properties']['message'];
-        }, {});
-        var err = {
-            msg: text,
-            fields_with_error: Object.keys(errors),
-            fields: fields,
-            type_error: "validator"
+User.createIndexes()
 
+module.exports.addOneUser = async function (user, callback) {
+    try {
+        var new_user = new User(user);
+        var errors = new_user.validateSync();
+        if (errors) {
+            errors = errors['errors'];
+            var text = Object.keys(errors).map((e) => {
+                return errors[e]['properties']['message'];
+            }).join(' ');
+            var fields = _.transform(Object.keys(errors), function (result, value) {
+                result[value] = errors[value]['properties']['message'];
+            }, {});
+            var err = {
+                msg: text,
+                fields_with_error: Object.keys(errors),
+                fields: fields,
+                type_error: "validator"
+            };
+            callback(err);
+        } else {
+            await new_user.save();
+            callback(null, new_user.toObject());
         }
-        callback(err)
-    } else {
-        new_user.save()
-        callback(null, new_user.toObject())
+    } catch (error) {
+        if (error.code === 11000) { // Erreur de duplicité
+            var field = Object.keys(error.keyValue)[0];
+            var err = {
+                msg: `Duplicate key error: ${field} must be unique.`,
+                fields_with_error: [field],
+                fields: { [field]: `The ${field} is already taken.` },
+                type_error: "duplicate"
+            };
+            callback(err);
+        } else {
+            callback(error); // Autres erreurs
+        }
     }
-}
+};
 
-module.exports.addManyUsers = function (users, callback) {
-    var errors = []
+module.exports.addManyUsers = async function (users, callback) {
+    var errors = [];
+    
+    // Vérifier les erreurs de validation
     for (var i = 0; i < users.length; i++) {
-        var user = users[i]
-        var new_user = new User(user)
-        var error = new_user.validateSync()
+        var user = users[i];
+        var new_user = new User(user);
+        var error = new_user.validateSync();
         if (error) {
-            error = error['errors']
+            error = error['errors'];
             var text = Object.keys(error).map((e) => {
-                return error[e]['properties']['message']
-            }).join(' ')
+                return error[e]['properties']['message'];
+            }).join(' ');
             var fields = _.transform(Object.keys(error), function (result, value) {
                 result[value] = error[value]['properties']['message'];
             }, {});
@@ -51,21 +69,36 @@ module.exports.addManyUsers = function (users, callback) {
                 fields: fields,
                 index: i,
                 type_error: "validator"
-            })
+            });
         }
     }
+    
     if (errors.length > 0) {
-        callback(errors)
+        callback(errors);
+    } else {
+        try {
+            // Tenter d'insérer les utilisateurs
+            const data = await User.insertMany(users, { ordered: false });
+            callback(null, data);
+        } catch (error) {
+            if (error.code === 11000) { // Erreur de duplicité
+                const duplicateErrors = error.writeErrors.map(err => {
+                    const field = err.err.errmsg.split(" dup key: { ")[1].split(':')[0].trim()
+                    return {
+                        msg: `Duplicate key error: ${field} must be unique.`,
+                        fields_with_error: [field],
+                        fields: { [field]: `The ${field} is already taken.` },
+                        index: err.index,
+                        type_error: "duplicate"
+                    };
+                });
+                callback(duplicateErrors);
+            } else {
+                callback(error); // Autres erreurs
+            }
+        }
     }
-    else {
-        User.insertMany(users).then(data => callback(null, data))
-            .catch(e => callback(e));
-        //  User.insertMany(users, callback)
-        //callback(null, users)
-        /* .then(data => callback(null, data))
-        .catch(e => callback(e)); */
-    }
-}
+};
 
 module.exports.findOneUser = function (user_id, callback) {
     if (user_id && mongoose.isValidObjectId(user_id)) {
@@ -93,7 +126,7 @@ module.exports.findManyUsers = function (users_id, callback) {
         users_id = users_id.map((e) => { return new ObjectId(e) })
         User.find({ _id: users_id }).then((value) => {
             try {
-                if (value) {
+                if (value && Array.isArray(value) &&value.length>0) {
                     callback(null, value);
                 } else {
                     callback({ msg: "Aucun utilisateur trouvé.", type_error: "no-found" });
@@ -119,53 +152,28 @@ module.exports.findManyUsers = function (users_id, callback) {
 }
 
 module.exports.updateOneUser = function (user_id, update, callback) {
-    if (user_id && mongoose.isValidObjectId(user_id)) {
+    if (user_id && mongoose.isValidObjectId(user_id)) {  
         User.findByIdAndUpdate(new ObjectId(user_id), update, { returnDocument: 'after', runValidators: true }).then((value) => {
             try {
-                callback(null, value.toObject())
+                if (value){
+                    callback(null, value.toObject())}
+                else{
+                    callback({ msg: "Utilisateur non trouvé.", type_error: "no-found" })}
             } catch (e) {
-                console.log(e)
-                callback(e)
+                callback({ msg: "Problème base de données", type_error: "error-mongo" })
             }
         }).catch((errors) => {
-            errors = errors['errors']
-            var text = Object.keys(errors).map((e) => {
-                return errors[e]['properties']['message']
-            }).join(' ')
-            var fields = _.transform(Object.keys(errors), function (result, value) {
-                result[value] = errors[value]['properties']['message'];
-            }, {});
-            var err = {
-                msg: text,
-                fields_with_error: Object.keys(errors),
-                fields: fields,
-                type_error: "validator"
-            }
-            callback(err)
-        })
-    }
-    else {
-        callback({ msg: "Id invalide.", type_error: 'no-valid' })
-    }
-}
-
-
-module.exports.updateManyUsers = function (users_id, update, callback) {
-    if(update && typeof update === "object" && !Array.isArray(update)){
-        
-        if (users_id && Array.isArray(users_id) && users_id.length > 0) {
-        
-            users_id = users_id.map((e) => { return new ObjectId(e) })
-            User.updateMany({ _id: users_id }, update, { runValidators: true }).then((value) => {
-                try {
-                    callback(null, value)
-                } catch (e) {
-                    console.log(e)
-                    callback(e)
-                }
-            }).catch((errors) => {
+            if (errors.code === 11000) { // Erreur de duplicité
+                var field = Object.keys(errors.keyPattern)[0];
+                var err = {
+                    msg: `Duplicate key error: ${field} must be unique.`,
+                    fields_with_error: [field],
+                    fields: { [field]: `The ${field} is already taken.` },
+                    type_error: "duplicate"
+                };
+                callback(err);
+            }else{
                 errors = errors['errors']
-                console.log('errors :', errors)
                 var text = Object.keys(errors).map((e) => {
                     return errors[e]['properties']['message']
                 }).join(' ')
@@ -179,18 +187,69 @@ module.exports.updateManyUsers = function (users_id, update, callback) {
                     type_error: "validator"
                 }
                 callback(err)
+            }
+        })
+    }else {
+        callback({ msg: "Id invalide.", type_error: 'no-valid' })
+    }
+}
+
+
+module.exports.updateManyUsers = function (users_id, update, callback) {0
+    if(users_id && Array.isArray(users_id) && users_id.length > 0 && users_id.filter((e) => { return mongoose.isValidObjectId(e)}).length == users_id.length){
+        if (users_id && Array.isArray(users_id) && users_id.length > 0) {
+            users_id = users_id.map((e) => { return new ObjectId(e) })
+            User.updateMany({ _id: users_id }, update, { runValidators: true }).then((value) => {
+                try {
+                    if (value && value.modifiedCount !== 0)
+                        callback(null, value)
+                    else
+                        callback({ msg: "Utilisateur non trouvé.", type_error: "no-found" });
+                } catch (e) {
+                    callback(e);
+                }
+            }).catch((errors) => {
+                if (errors.code === 11000) { // Erreur de duplicité
+                    var field = Object.keys(errors.keyPattern)[0]
+                    var err = {
+                        msg: `Duplicate key error: ${field} must be unique.`,
+                        fields_with_error: [field],
+                        fields: { [field]: `The ${field} is already taken.` },
+                        type_error: "duplicate"
+                    };
+                    callback(err);
+                }else{
+                    errors = errors['errors']
+                    var text = Object.keys(errors).map((e) => {
+                        return errors[e]['properties']['message']
+                    }).join(' ')
+                    var fields = _.transform(Object.keys(errors), function (result, value) {
+                        result[value] = errors[value]['properties']['message'];
+                    }, {});
+                    var err = {
+                        msg: text,
+                        fields_with_error: Object.keys(errors),
+                        fields: fields,
+                        type_error: "validator"
+                    }
+                    callback(err)
+                }
             })
         }
         else {
             callback({ msg: "Id invalide.", type_error: 'no-valid' })
         }
+    }else if (users_id && Array.isArray(users_id) && users_id.length >  0 && users_id.filter((e) => { return mongoose.isValidObjectId(e)}).length != users_id.length) {
+        callback({ msg: "Tableau non conforme plusieurs éléments ne sont pas des ObjectId.", type_error: 'no-valid', fields: users_id.filter((e) => { return !mongoose.isValidObjectId(e)}) });
     }else{
+        console.log('ok')
         callback({ msg: "Le body doit etre un objet.", type_error: 'no-valid' })
     }
 }
 
 module.exports.deleteOneUser = function (user_id, callback) {
     if (user_id && mongoose.isValidObjectId(user_id)) {
+        
         User.findByIdAndDelete(user_id).then((value) => {
             try {
                 if (value)
@@ -211,16 +270,28 @@ module.exports.deleteOneUser = function (user_id, callback) {
     }
 }
 
-module.exports.deleteManyUsers = function(users_id, callback) {
-    if (users_id && Array.isArray(users_id) && users_id.length > 0) {
-        users_id = users_id.map((e) => { return new ObjectId(e)})
-        User.deleteMany({_id: users_id}).then((value) => {
-            callback(null, value)
+module.exports.deleteManyUsers = function (users_id, callback) {
+    if (users_id && Array.isArray(users_id) && users_id.length > 0 && users_id.filter((e) => { return mongoose.isValidObjectId(e) }).length == users_id.length) {
+        users_id = users_id.map((e) => { return new ObjectId(e) })
+        User.deleteMany({ _id: users_id }).then((value) => {
+            if (value && value.deletedCount !==0) {
+                    callback(null, value);
+                } else {
+                    
+                    callback({ msg: "Aucun utilisateur trouvé.", type_error: "no-found" });
+                }
         }).catch((err) => {
-            callback({ msg: "Erreur mongo suppression.", type_error: "error-mongo" }); 
+            callback({ msg: "Id inconnu", type_error: "no-found" });
         })
     }
+    else if (users_id && Array.isArray(users_id) && users_id.length > 0 && users_id.filter((e) => { return mongoose.isValidObjectId(e) }).length != users_id.length) {
+        callback({ msg: "Tableau non conforme plusieurs éléments ne sont pas des ObjectId.", type_error: 'no-valid', fields: users_id.filter((e) => { return !mongoose.isValidObjectId(e) }) });
+    }
+    else if (users_id && !Array.isArray(users_id)) {
+        callback({ msg: "L'argement n'est pas un tableau.", type_error: 'no-valid' });
+
+    }
     else {
-        callback({ msg: "Tableau d'id invalide.", type_error: 'no-valid' })
+        callback({ msg: "Tableau non conforme.", type_error: 'no-valid' });
     }
 }
